Introduce setupRandomColony
