+++ test/meta-colony.js
@@ -1,6 +1,6 @@
 /* globals artifacts */
 import { INITIAL_FUNDING, DELIVERABLE_HASH } from "../helpers/constants";
-import { checkErrorRevert, getTokenArgs } from "../helpers/test-helper";
+import { checkErrorRevert } from "../helpers/test-helper";
 import {
   fundColonyWithTokens,
   setupFundedTask,
@@ -8,14 +8,13 @@ import {
   executeSignedTaskChange,
   makeTask,
   setupColonyNetwork,
-  setupMetaColonyWithLockedCLNYToken
+  setupMetaColonyWithLockedCLNYToken,
+  setupRandomColony
 } from "../helpers/test-data-generator";
 
-const IColony = artifacts.require("IColony");
 const ERC20ExtendedToken = artifacts.require("ERC20ExtendedToken");
 
 contract("Meta Colony", accounts => {
-  let TOKEN_ARGS;
   const MANAGER = accounts[0];
   const OTHER_ACCOUNT = accounts[1];
   const WORKER = accounts[2];
@@ -289,11 +288,8 @@ contract("Meta Colony", accounts => {
 
   describe("when adding domains in a regular colony", () => {
     beforeEach(async () => {
-      TOKEN_ARGS = getTokenArgs();
-      const newToken = await ERC20ExtendedToken.new(...TOKEN_ARGS);
-      const { logs } = await colonyNetwork.createColony(newToken.address);
-      const { colonyAddress } = logs[0].args;
-      colony = await IColony.at(colonyAddress);
+      colony = await setupRandomColony(colonyNetwork);
+
       const tokenAddress = await colony.getToken();
       token = await ERC20ExtendedToken.at(tokenAddress);
     });
@@ -355,12 +351,10 @@ contract("Meta Colony", accounts => {
 
   describe("when setting domain and skill on task", () => {
     beforeEach(async () => {
-      TOKEN_ARGS = getTokenArgs();
-      token = await ERC20ExtendedToken.new(...TOKEN_ARGS);
-      const { logs } = await colonyNetwork.createColony(token.address);
-      const { colonyAddress } = logs[0].args;
-      await token.setOwner(colonyAddress);
-      colony = await IColony.at(colonyAddress);
+      colony = await setupRandomColony(colonyNetwork);
+
+      const tokenAddress = await colony.getToken();
+      token = await ERC20ExtendedToken.at(tokenAddress);
     });
 
     it("should be able to set domain on task", async () => {
